Move initialisation into AppComponents where it belongs
